Improve projector cube mode render
No long hides water and weather. Credit to Mekanism for their useful render code!

+++ src/main/java/dev/su5ed/mffs/blockentity/FortronBlockEntity.java
@@ -56,7 +56,7 @@ public abstract class FortronBlockEntity extends InventoryBlockEntity implements
         this.frequencySlot = addSlot("frequency", InventorySlot.Mode.NONE, stack -> stack.getItem() instanceof Card);
     }
 
-    public long getAnimation() {
+    public int getAnimation() {
         return this.animation;
     }
